requirements.txt
flask==2.2.2
marshmallow==3.19.0
flask_sqlalchemy==3.0.2
flask_migrate==4.0.2
psycopg2==2.9.5